[WIP] CollectPointsRays
Removed old code for the rays but left the one where the old rays to
calculated if this shall be used for a comparison some time.

Only thing left to do: Old ray distances have been divided by screen
width/height, this is not done with the new rays. Maybe this is not an
issue.