GUIApp: Changed reward computation to use a NumPy array that stores the rewards
- Created the RewardElement class. Classes that give out rewards inherit from RewardElement and provide
a reward_template dict, from which a reward_mapping dict is created. That can be used to index the
reward_array of that class, a NumPy array containing initially only zeros, and once the respective reward
is given out, it contains a 1 at the respective index.
- Comparing the reward arrays between each step is computationally much faster, than comparing the reward dicts
as done before. Additionally, as the reward array is flat, it only needs copy, compared to the previous version
which used deepcopy, as the dict was nested

Signed-off-by: Patrick Deubel <33164686+pdeubel@users.noreply.github.com>